--- websdk-java-core/src/main/java/cn/xfyun/basic/ObjectMapperFactory.java
@@ -1,23 +1,0 @@
-package cn.xfyun.basic;
-
-import com.fasterxml.jackson.databind.DeserializationFeature;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
-
-/**
- * @author: rblu2
- * @create: 2024-12-23 11:27
- **/
-public class ObjectMapperFactory {
-
-    private static final ObjectMapper objectMapper = new ObjectMapper();
-
-    static {
-        objectMapper.registerModule(new JavaTimeModule());
-        objectMapper.configure(DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES, false);
-    }
-    
-    public static ObjectMapper instance() {
-        return objectMapper;
-    }
-}
